backend/requirements.txt
bcrypt
email-validator
fastapi
motor
pydantic
pyjwt
pymongo
uvicorn